Revert "Dt bugs"
