+++ App.tsx
@@ -1,5 +1,27 @@
-import RootNavigator from "./navigation/RootNavigator";
+
+import { StatusBar } from 'expo-status-bar';
+import { NavigationContainer } from '@react-navigation/native';
+import { StyleSheet, View } from 'react-native';
+import { useFonts, Roboto_400Regular, Roboto_700Bold } from '@expo-google-fonts/roboto';
+import AppLoading from 'expo-app-loading';
+import React from 'react';
+import { createStackNavigator } from '@react-navigation/stack';
+import SignUpScreen from './components/SignUpScreen';
+
+const Stack = createStackNavigator();
 
 export default function App() {
-  return <RootNavigator />;
+
+
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="SignUp">
+        <Stack.Screen name="SignUp" component={SignUpScreen} options={{ headerShown: false }} />
+      </Stack.Navigator>
+      <StatusBar style="auto" />
+    </NavigationContainer>
+  );
 }
+
+
+
