+++ components/ActivePostsPage.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, Text, View, Image, ScrollView } from 'react-native';
 import React from "react"
-import FarmerPost from "../jobcomponents/FarmerPost"
+import FarmerPost from "./jobcomponents/FarmerPost"
 
 
 const ActivePostsPage = () => {
