+++ components/EmailPage.tsx
