+++ components/SignUpScreen.tsx
@@ -2,10 +2,11 @@ import React, { useState } from "react";
 import { View, Text, TextInput, TouchableOpacity, Image, StyleSheet, Alert} from "react-native";
 import * as ImagePicker from "expo-image-picker";
 import { Button } from "react-native-paper";
-import { supabase } from "../../lib/supabase";
+import { supabase } from "../lib/supabase";
 
 
-const SignUpScreen: React.FC<{ navigation: any }> = ({ navigation }) => {
+
+const SignUpScreen = () => {
   const [firstName, setFirstName] = useState("");
   const [lastName, setLastName] = useState("");
   const [email, setEmail] = useState("");
@@ -30,10 +31,6 @@ const SignUpScreen: React.FC<{ navigation: any }> = ({ navigation }) => {
 
   
 async function signUp() {
-  if(!email || !password || !firstName || !lastName || !role) {
-    Alert.alert("Error", "Please fill in all fields.");
-    return;
-  }
   setLoading(true);
   console.log("🔄 Signing up user...");
 
@@ -64,14 +61,15 @@ async function signUp() {
     setLoading(false);
     return;
   }
-console.log("auth.uid() at insert:", supabase.auth.getSession());
 
   console.log("📤 Inserting user into custom users table...");
   const { error: insertError } = await supabase.from("accounts").insert([
     {
-      email: email,
+      uuid: userId,
+      email,
       first_name: firstName,
       last_name: lastName,
+      password,
       role
     },
   ]);
@@ -93,7 +91,7 @@ console.log("auth.uid() at insert:", supabase.auth.getSession());
 
   return (
     <View style={styles.container}>
-      <Image source={require("../../assets/logo.png")} style={styles.logo} />
+      <Image source={require("../assets/logo.png")} style={styles.logo} />
 
       <View style={styles.card}>
         <Text style={styles.title}>Sign Up</Text>
@@ -122,10 +120,10 @@ console.log("auth.uid() at insert:", supabase.auth.getSession());
             onChangeText={setPassword}
           />
           <TouchableOpacity
-            onPress={() => setShowPassword((prev) => !prev)}
+            onPress = {() => setShowPassword(prev => !prev)}
             style={{ position: "absolute", right: 10, top: 12 }}
           >
-            <Text> {showPassword ? "🔒" : "👁️"} </Text>
+            <Text> {showPassword? '🔒' : '👁️'} </Text>
           </TouchableOpacity>
         </View>
         <TextInput
@@ -175,13 +173,6 @@ console.log("auth.uid() at insert:", supabase.auth.getSession());
         >
           Confirm
         </Button>
-        <TouchableOpacity onPress={() => navigation.navigate("Login")}>
-          <Text
-            style={{ marginTop: 15, textAlign: "center", color: "#3F6939" }}
-          >
-            Already have an account? Log In
-          </Text>
-        </TouchableOpacity>
       </View>
     </View>
   );
