+++ components/Tabs.tsx
@@ -1,9 +1,9 @@
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs"
 import { StyleSheet, Text, View, Image } from 'react-native';
 import React from "react"
-import JobsPage from "./farmer/JobsPage"
-import AddPage from "./farmer/AddPage"
-import EmailPage from "./EmailPage"
+import JobsPage from "./../components/JobsPage"
+import AddPage from "./../components/AddPage"
+import EmailPage from "./../components/EmailPage"
 
 const Tab = createBottomTabNavigator();
