+++ lib/supabase.ts
@@ -1,15 +1,9 @@
 import { AppState } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { createClient } from '@supabase/supabase-js';
-import { useState, useEffect } from 'react';
-import { User } from '@supabase/supabase-js';
 
-const url = process.env.EXPO_PUBLIC_SUPABASE_URL!;
-const key = process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY!;
-
-if (!url || !key) {
-  throw new Error('Supabase URL or ANON KEY is not defined in environment variables.');
-}
+const url = process.env.EXPO_PUBLIC_SUPABASE_URL;
+const key = process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY;
 
 export const supabase = createClient(url, key, {
   auth: {
@@ -28,5 +22,3 @@ AppState.addEventListener('change', (state) => {
     supabase.auth.stopAutoRefresh();
   }
 });
-
-
